+++ src/main/java/roomescape/exception/login/AccessDeniedException.java
@@ -1,7 +1,7 @@
 package roomescape.exception.login;
 
 import org.springframework.http.HttpStatus;
-import roomescape.exception.RoomescapeException;
+import roomescape.exception.common.RoomescapeException;
 
 public class AccessDeniedException extends RoomescapeException {
     public AccessDeniedException() {
