+++ src/main/java/roomescape/exception/login/ExpiredTokenException.java
@@ -1,7 +1,7 @@
 package roomescape.exception.login;
 
 import org.springframework.http.HttpStatus;
-import roomescape.exception.RoomescapeException;
+import roomescape.exception.common.RoomescapeException;
 
 public class ExpiredTokenException extends RoomescapeException {
     public ExpiredTokenException() {
