+++ src/main/java/roomescape/exception/member/DuplicatedMemberEmailException.java
@@ -1,7 +1,7 @@
 package roomescape.exception.member;
 
 import org.springframework.http.HttpStatus;
-import roomescape.exception.RoomescapeException;
+import roomescape.exception.common.RoomescapeException;
 
 public class DuplicatedMemberEmailException extends RoomescapeException {
     public DuplicatedMemberEmailException() {
