+++ src/main/java/roomescape/exception/member/InvalidMemberPasswordLengthException.java
@@ -1,7 +1,7 @@
 package roomescape.exception.member;
 
 import org.springframework.http.HttpStatus;
-import roomescape.exception.RoomescapeException;
+import roomescape.exception.common.RoomescapeException;
 
 public class InvalidMemberPasswordLengthException extends RoomescapeException {
     public InvalidMemberPasswordLengthException() {
