+++ src/main/java/roomescape/exception/theme/InvalidThemeNameLengthException.java
@@ -1,7 +1,7 @@
 package roomescape.exception.theme;
 
 import org.springframework.http.HttpStatus;
-import roomescape.exception.RoomescapeException;
+import roomescape.exception.common.RoomescapeException;
 
 public class InvalidThemeNameLengthException extends RoomescapeException {
     public InvalidThemeNameLengthException() {
