+++ src/main/java/roomescape/exception/time/ReservationReferencedTimeException.java
@@ -1,7 +1,7 @@
 package roomescape.exception.time;
 
 import org.springframework.http.HttpStatus;
-import roomescape.exception.RoomescapeException;
+import roomescape.exception.common.RoomescapeException;
 
 public class ReservationReferencedTimeException extends RoomescapeException {
     public ReservationReferencedTimeException() {
